# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Gateway universel : utilise tes outils habituels (Claude CLI, Cursor, Aider) avec des modèles moins chers en arrière-plan (DeepSeek, Groq, Mistral). Garde ton workflow. Réduis ta facture API de 90%.",
        "title": "Veil API",
        "contact": {
            "name": "Support Veil",
            "email": "support@veil.dev"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/billing/plan": {
            "get": {
                "security": [
                    {
                        "DashboardAuth": []
                    }
                ],
                "description": "Retourne le plan souscrit, son prix, le quota mensuel de tokens et le nombre de clés API autorisées.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Plan actuel",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/billing.PlanRecord"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/billing.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/billing.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/billing/upgrade": {
            "post": {
                "security": [
                    {
                        "DashboardAuth": []
                    }
                ],
                "description": "Génère une URL Stripe Checkout pour upgrader vers starter, pro ou team. Redirige l'utilisateur vers cette URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Mettre à niveau le plan",
                "parameters": [
                    {
                        "description": "Plan cible",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/billing.UpgradePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL de paiement Stripe",
                        "schema": {
                            "$ref": "#/definitions/billing.UpgradePlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/billing.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/billing.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/billing.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/keys": {
            "get": {
                "security": [
                    {
                        "DashboardAuth": []
                    }
                ],
                "description": "Retourne toutes les clés actives. La valeur brute n'est jamais exposée ici.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-keys"
                ],
                "summary": "Lister les clés API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/keys.ListKeysResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/keys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/keys.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "DashboardAuth": []
                    }
                ],
                "description": "Génère une nouvelle clé vl_live_xxx. La valeur brute n'est retournée qu'une seule fois.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-keys"
                ],
                "summary": "Créer une clé API",
                "parameters": [
                    {
                        "description": "Nom de la clé",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/keys.CreateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Clé créée — conserver la valeur 'key', elle ne sera plus visible",
                        "schema": {
                            "$ref": "#/definitions/keys.CreateKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/keys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/keys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/keys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/keys/{id}": {
            "delete": {
                "security": [
                    {
                        "DashboardAuth": []
                    }
                ],
                "description": "Supprime définitivement la clé. Les requêtes utilisant cette clé seront immédiatement rejetées.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-keys"
                ],
                "summary": "Révoquer une clé API",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"550e8400-e29b-41d4-a716-446655440000\"",
                        "description": "ID de la clé",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Clé révoquée"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/keys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/keys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/keys.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "billing.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
        "billing.PlanRecord": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "free"
                },
                "max_api_keys": {
                    "type": "integer",
                    "example": 1
                },
                "price_usd": {
                    "type": "number",
                    "example": 0
                },
                "token_quota": {
                    "type": "integer",
                    "example": 100000
                }
            }
        },
        "billing.UpgradePlanRequest": {
            "type": "object",
            "properties": {
                "plan": {
                    "type": "string",
                    "example": "starter"
                }
            }
        },
        "billing.UpgradePlanResponse": {
            "type": "object",
            "properties": {
                "checkout_url": {
                    "type": "string",
                    "example": "https://checkout.stripe.com/pay/cs_test_xxx"
                }
            }
        },
...
        "keys.CreateKeyRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "cursor-work"
                }
            }
        },
        "keys.CreateKeyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-05-15T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "key": {
                    "type": "string",
                    "example": "vl_live_aB3xK9mN2pQrS7tU1vW4yZ8cD6eF0gH5jI"
                },
                "name": {
                    "type": "string",
                    "example": "cursor-work"
                }
            }
        },
        "keys.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "key not found"
                }
            }
        },
        "keys.KeyItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-05-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "last_used": {
                    "type": "string",
                    "example": "2026-05-15T10:30:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "cursor-work"
                }
            }
        },
        "keys.ListKeysResponse": {
            "type": "object",
            "properties": {
                "keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/keys.KeyItem"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "DashboardAuth": {
            "description": "JWT Dashboard — pour les endpoints /api/*",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "VeilAPIKey": {
            "description": "Clé Veil vl_live_xxx — pour les endpoints /v1/*",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}